feat: move sepc_filter from storage to spec, use the YAML format file
