+++ custom_components/xiaomi_home/miot/miot_storage.py
@@ -719,113 +719,6 @@ class MIoTCert:
         return binascii.hexlify(sha1_hash.finalize()).decode('utf-8')
 
 
-class SpecFilter:
-    """
-    MIoT-Spec-V2 filter for entity conversion.
-    """
-    SPEC_FILTER_FILE = 'specs/spec_filter.json'
-    _main_loop: asyncio.AbstractEventLoop
-    _data: dict[str, dict[str, set]]
-    _cache: Optional[dict]
-
-    def __init__(self, loop: Optional[asyncio.AbstractEventLoop]) -> None:
-        self._main_loop = loop or asyncio.get_event_loop()
-        self._data = None
-        self._cache = None
-
-    async def init_async(self) -> None:
-        if isinstance(self._data, dict):
-            return
-        filter_data = None
-        self._data = {}
-        try:
-            filter_data = await self._main_loop.run_in_executor(
-                None, load_json_file,
-                os.path.join(
-                    os.path.dirname(os.path.abspath(__file__)),
-                    self.SPEC_FILTER_FILE))
-        except Exception as err:  # pylint: disable=broad-exception-caught
-            _LOGGER.error('spec filter, load file error, %s', err)
-            return
-        if not isinstance(filter_data, dict):
-            _LOGGER.error('spec filter, invalid spec filter content')
-            return
-        for values in list(filter_data.values()):
-            if not isinstance(values, dict):
-                _LOGGER.error('spec filter, invalid spec filter data')
-                return
-            for value in values.values():
-                if not isinstance(value, list):
-                    _LOGGER.error('spec filter, invalid spec filter rules')
-                    return
-
-        self._data = filter_data
-
-    async def deinit_async(self) -> None:
-        self._cache = None
-        self._data = None
-
-    async def set_spec_spec(self, urn_key: str) -> None:
-        """MUST call init_async() first."""
-        if not self._data:
-            return
-        self._cache = self._data.get(urn_key, None)
-
-    def filter_service(self, siid: int) -> bool:
-        """Filter service by siid.
-        MUST call init_async() and set_spec_spec() first."""
-        if (
-            self._cache
-            and 'services' in self._cache
-            and (
-                str(siid) in self._cache['services']
-                or '*' in self._cache['services'])
-        ):
-            return True
-
-        return False
-
-    def filter_property(self, siid: int, piid: int) -> bool:
-        """Filter property by piid.
-        MUST call init_async() and set_spec_spec() first."""
-        if (
-            self._cache
-            and 'properties' in self._cache
-            and (
-                f'{siid}.{piid}' in self._cache['properties']
-                or f'{siid}.*' in self._cache['properties'])
-        ):
-            return True
-        return False
-
-    def filter_event(self, siid: int, eiid: int) -> bool:
-        """Filter event by eiid.
-        MUST call init_async() and set_spec_spec() first."""
-        if (
-            self._cache
-            and 'events' in self._cache
-            and (
-                f'{siid}.{eiid}' in self._cache['events']
-                or f'{siid}.*' in self._cache['events']
-            )
-        ):
-            return True
-        return False
-
-    def filter_action(self, siid: int, aiid: int) -> bool:
-        """"Filter action by aiid.
-        MUST call init_async() and set_spec_spec() first."""
-        if (
-            self._cache
-            and 'actions' in self._cache
-            and (
-                f'{siid}.{aiid}' in self._cache['actions']
-                or f'{siid}.*' in self._cache['actions'])
-        ):
-            return True
-        return False
-
-
 class DeviceManufacturer:
     """Device manufacturer."""
     DOMAIN: str = 'miot_specs'
