+++ custom_components/xiaomi_home/miot/specs/bool_trans.yaml
@@ -1,0 +1,231 @@
+data:
+  urn:miot-spec-v2:property:air-cooler:000000EB: open_close
+  urn:miot-spec-v2:property:alarm:00000012: open_close
+  urn:miot-spec-v2:property:anion:00000025: open_close
+  urn:miot-spec-v2:property:anti-fake:00000130: yes_no
+  urn:miot-spec-v2:property:arrhythmia:000000B4: yes_no
+  urn:miot-spec-v2:property:auto-cleanup:00000124: open_close
+  urn:miot-spec-v2:property:auto-deodorization:00000125: open_close
+  urn:miot-spec-v2:property:auto-keep-warm:0000002B: open_close
+  urn:miot-spec-v2:property:automatic-feeding:000000F0: open_close
+  urn:miot-spec-v2:property:blow:000000CD: open_close
+  urn:miot-spec-v2:property:card-insertion-state:00000106: yes_no
+  urn:miot-spec-v2:property:contact-state:0000007C: contact_state
+  urn:miot-spec-v2:property:current-physical-control-lock:00000099: open_close
+  urn:miot-spec-v2:property:delay:0000014F: yes_no
+  urn:miot-spec-v2:property:deodorization:000000C6: open_close
+  urn:miot-spec-v2:property:dns-auto-mode:000000DC: open_close
+  urn:miot-spec-v2:property:driving-status:000000B9: yes_no
+  urn:miot-spec-v2:property:dryer:00000027: open_close
+  urn:miot-spec-v2:property:eco:00000024: open_close
+  urn:miot-spec-v2:property:glimmer-full-color:00000089: open_close
+  urn:miot-spec-v2:property:guard-mode:000000B6: open_close
+  urn:miot-spec-v2:property:heater:00000026: open_close
+  urn:miot-spec-v2:property:heating:000000C7: open_close
+  urn:miot-spec-v2:property:horizontal-swing:00000017: open_close
+  urn:miot-spec-v2:property:hot-water-recirculation:0000011C: open_close
+  urn:miot-spec-v2:property:image-distortion-correction:0000010F: open_close
+  urn:miot-spec-v2:property:local-storage:0000011E: yes_no
+  urn:miot-spec-v2:property:motion-detection:00000056: open_close
+  urn:miot-spec-v2:property:motion-state:0000007D: motion_state
+  urn:miot-spec-v2:property:motion-tracking:0000008A: open_close
+  urn:miot-spec-v2:property:motor-reverse:00000072: yes_no
+  urn:miot-spec-v2:property:mute:00000040: open_close
+  urn:miot-spec-v2:property:off-delay:00000053: open_close
+  urn:miot-spec-v2:property:on:00000006: open_close
+  urn:miot-spec-v2:property:physical-controls-locked:0000001D: open_close
+  urn:miot-spec-v2:property:plasma:00000132: yes_no
+  urn:miot-spec-v2:property:preheat:00000103: open_close
+  urn:miot-spec-v2:property:seating-state:000000B8: yes_no
+  urn:miot-spec-v2:property:silent-execution:000000FB: yes_no
+  urn:miot-spec-v2:property:sleep-aid-mode:0000010B: open_close
+  urn:miot-spec-v2:property:sleep-mode:00000028: open_close
+  urn:miot-spec-v2:property:snore-state:0000012A: yes_no
+  urn:miot-spec-v2:property:soft-wind:000000CF: open_close
+  urn:miot-spec-v2:property:speed-control:000000E8: open_close
+  urn:miot-spec-v2:property:submersion-state:0000007E: yes_no
+  urn:miot-spec-v2:property:time-watermark:00000087: open_close
+  urn:miot-spec-v2:property:un-straight-blowing:00000100: open_close
+  urn:miot-spec-v2:property:uv:00000029: open_close
+  urn:miot-spec-v2:property:valve-switch:000000FE: open_close
+  urn:miot-spec-v2:property:ventilation:000000CE: open_close
+  urn:miot-spec-v2:property:vertical-swing:00000018: open_close
+  urn:miot-spec-v2:property:wake-up-mode:00000107: open_close
+  urn:miot-spec-v2:property:water-pump:000000F2: open_close
+  urn:miot-spec-v2:property:watering:000000CC: open_close
+  urn:miot-spec-v2:property:wdr-mode:00000088: open_close
+  urn:miot-spec-v2:property:wet:0000002A: open_close
+  urn:miot-spec-v2:property:wifi-band-combine:000000E0: open_close
+  urn:miot-spec-v2:property:wifi-ssid-hidden:000000E3: yes_no
+  urn:miot-spec-v2:property:wind-reverse:00000117: yes_no
+translate:
+  default:
+    de:
+      true: Wahr
+      false: Falsch
+    en:
+      true: True
+      false: False
+    es:
+      true: Verdadero
+      false: Falso
+    fr:
+      true: Vrai
+      false: Faux
+    ja:
+      true: 真
+      false: 偽
+    nl:
+      true: True
+      false: False
+    pt:
+      true: True
+      false: False
+    pt-BR:
+      true: True
+      false: False
+    ru:
+      true: Истина
+      false: Ложь
+    zh-Hans:
+      true: 真
+      false: 假
+    zh-Hant:
+      true: 真
+      false: 假
+  open_close:
+    de:
+      true: Öffnen
+      false: Schließen
+    en:
+      true: Open
+      false: Close
+    es:
+      true: Abierto
+      false: Cerrado
+    fr:
+      true: Ouvert
+      false: Fermer
+    ja:
+      true: 開く
+      false: 閉じる
+    nl:
+      true: Open
+      false: Dicht
+    pt:
+      true: Aberto
+      false: Fechado
+    pt-BR:
+      true: Aberto
+      false: Fechado
+    ru:
+      true: Открыть
+      false: Закрыть
+    zh-Hans:
+      true: 开启
+      false: 关闭
+    zh-Hant:
+      true: 開啟
+      false: 關閉
+  yes_no:
+    de:
+      true: Ja
+      false: Nein
+    en:
+      true: Yes
+      false: No
+    es:
+      true: Sí
+      false: No
+    fr:
+      true: Oui
+      false: Non
+    ja:
+      true: はい
+      false: いいえ
+    nl:
+      true: Ja
+      false: Nee
+    pt:
+      true: Sim
+      false: Não
+    pt-BR:
+      true: Sim
+      false: Não
+    ru:
+      true: Да
+      false: Нет
+    zh-Hans:
+      true: 是
+      false: 否
+    zh-Hant:
+      true: 是
+      false: 否
+  motion_state:
+    de:
+      true: Bewegung erkannt
+      false: Keine Bewegung erkannt
+    en:
+      true: Motion Detected
+      false: No Motion Detected
+    es:
+      true: Movimiento detectado
+      false: No se detecta movimiento
+    fr:
+      true: Mouvement détecté
+      false: Aucun mouvement détecté
+    ja:
+      true: 動きを検知
+      false: 動きが検出されません
+    nl:
+      true: Contact
+      false: Geen contact
+    pt:
+      true: Contato
+      false: Sem contato
+    pt-BR:
+      true: Contato
+      false: Sem contato
+    ru:
+      true: Обнаружено движение
+      false: Движение не обнаружено
+    zh-Hans:
+      true: 有人
+      false: 无人
+    zh-Hant:
+      true: 有人
+      false: 無人
+  contact_state:
+    de:
+      true: Kontakt
+      false: Kein Kontakt
+    en:
+      true: Contact
+      false: No Contact
+    es:
+      true: Contacto
+      false: Sin contacto
+    fr:
+      true: Contact
+      false: Pas de contact
+    ja:
+      true: 接触
+      false: 非接触
+    nl:
+      true: Contact
+      false: Geen contact
+    pt:
+      true: Contato
+      false: Sem contato
+    pt-BR:
+      true: Contato
+      false: Sem contato
+    ru:
+      true: Контакт
+      false: Нет контакта
+    zh-Hans:
+      true: 接触
+      false: 分离
+    zh-Hant:
+      true: 接觸
+      false: 分離
